feat: Updated the 'forceDevMode' setting
Problem:

There was not a way to explicitly enable and disable
dev mode with a single setting

Solution:

Update the logic so that aws.dev.forceDevMode takes
precedence over all other settings and will honor
whatever that value is in relation to if the user
is in dev mode.

Now, a user can have all their other dev settings
in place, but when they set 'aws.dev.forceDevMode: false'
they will not be in dev mode anymore.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>