authWebview: Various small changes (#3538)
* Adding visual content + css

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* authWebview: can toggle click service item

When clicking a service item, you can click it
again to close it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* throttle connection change event firing

We listen to multiple connection change events
and then trigger a full rerender of all states.

Due to listening to multiple events that tend
to trigger at the same time, we change events
and then trigger a full rerender of all states
multiple times.

SOLUTION:

throttle the event firing so that it only happens
with a certain delay.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: Don't need to use mapping for types

I assumed we needed a mapping to have values which
can be referenced and refactored easily.

But by explicitly setting these values in a type
we are able to get the same functionality.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: dont need to use mapping for certain values

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: Update service content text to match updated mocks

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* credentialsForm: edit credentials file button

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* delayed rendering for auth loading

Problem

In the UI a content window (cc, cw, ...) can have multiple
auth forms within it.

These auth forms need to communicate with the backend
so they can update to the latest state (check if the
auth is already connected)

Since we want the auth form to be independent from its parent
component we need to rely on the auth form to resolve
its own status when it is created.

The real problem is that one version of the auth form is
displayed to the user, but the UI flickers as it updates
the form to the latest state.

Solution

Have the auth form make all its backend calls, but have
the form emit an event to indicate once it is done loading.

We can then delay displaying the auth form in its parent
component and only show once it is fully ready.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* add content for code catalyst window

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix resizing bug

In a certain case when resizing, the servic content window would not
show when clicked. This moves some functions out of the 'computed'
property and in to the 'methods'.

TODO: Look more in to why

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* placeholder pictures in content

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* merge lock/unlocked items

In the webview on the left side we previously
had 2 sections for Unlocked and Locked features.

This commit merges them in to 1 section

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* codecatalyst getStartedCommand reauthenticate if needed

Before this change, if this command was called and
the Builder ID was expired, we would fail the
isConnectionOnboarded() call.

This change ensures the connection is valid before
proceeding.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* minor refactor

rearrange html elements to align
the left + right content veritically
and have the header on top

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix: copyright header

Remove the dates from the copyright header to
pass eslint

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix: don't reload webview when switching tabs

If the webview is open and the user switches
tabs, when they switch back to the auth webview
it will reload it.

This commit sets the setting to prevent the reload
from happening

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: move types from vue to ts file

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: move another types file to the root

This types file does not need to be in a 'backend'
folder. Instead it is moved to the root of the webview
code.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>