codewhisperer: Update tests to point to new sso signup command
Problem:

Before the new auth webview, CW used a quickpick wizard for
users to setup their required auth (sso, builderid). They used
a specific command and ensured they used it in their unit tests.

But now with the auth webview changes we will not be redirecting
them to the quickpick wizard anymore. Instead, we will redirect
users to the webview. This uses a different command name.

Solution:

Update the unit tests to point to the new vscode command name
that will be used.

Additionally, since the whole auth webview feature is not ready
we fallback to the old functionality if the user is not in
dev mode (an aws extension setting).

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>