fix: vscode command gets wrong arg
When a vscode command is called it is possible for
vscode to pass in a completely separate type of object
as the argument, instead of the expected auth object.

Eg: If I select a node in the explorer then call this command
it will pass in that node as context.

Solution:

In this case we do not need auth as an argument and
can instead define it inside the function.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>