test: regionSubmenu tests don't need to be cwl specific #3579
Problem:
The tests currently use language specific to cwl, implying the component is only used for that purpose. However, nothing in the design of the UI component is specific to cwl and can be used outside that context.

Solution:
Change the language to be more general, and avoid cwl terms like `group` or `logData`.