identityCenter: Add identity center form state
This adds the class which manages the state of
identity center in the context of the Explorer.

In the context of the Explorer, Identity Center
form behaviour is slightly different which is reflected
in some of the implementations of the class.

In short, IC in Explorer is not required to be
actively connected. Instead it is just used to
get the Credentials profile information. So we
do not care about the active IC connection in this
scenario

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>