identityCenter: Show different form title for Explorer
With Identity Center in the context of the Explorer
we do not care about the actual 'active' status
of an Identity Center connection.

Due to this we don't want to show anything regarding
a 'successful' connection in the form title.

Solution:

Have a conditional title that simply shows the name
if we set a certain Vue Prop and ignores the connected
stuff.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>