authWebview: UI buttons open auth webview (#3539)
* refactor: make AuthCommandDeclarations singleton

This makes this class a singleton and updates
existing use cases

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: Developer Tools 'Start' button opens webview

When the 'Start' node for CC or CW is selected
this will now have the webview open instead
of the builder id setup prompts

- still need to setup the start button to cause
  webview to open on specific service item

- Also added a fallback when not in dev mode to use
the old add connection functionality.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpenView: 'Start' from developer tools opens with service already selected

When the user clicks the 'Start' button the auth webview
page will be openend with the specific service already
selected.

Eg: Clicking 'Start' in CodeCatalyst opens up the webview
directly to CodeCatalyst

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* explorer: vscode command to show 'developer tools' root node

- Creates vscode commands that can eventually be executed
  and will show the root node (cc, cw, ..) in the developer
  tools tree view

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* button in form to open node

After the user has signed in with BuilderID
to CC or CW, we will now show a button which
can be clicked and will open the appropriate
node in the 'Developer Tools' view.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* fix: circular dependency

Problem:

The getStartedCommand is now used in multiple
places outside of the explorer. This was causing
a circular dependency.

Solution:

Move the getStartedCommand to the utils module

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: Resource Explorer credentials form button

Adds a button once the resource explorer credentials
are authenticated to open the Resource Explorer
view

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: Identity Center form button opens node

Now once CW is signed in with Identity Center we will
show a button that expands the CW node in the Developer
Tools menu

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: 'Add New Connection' quick pick opens webview

In the connection selection quick pick the last option
is 'Add New Connection' that previously led in to
another vscode wizard.

Now this instead opens the Add Connection webview.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* codewhisperer: Update tests to point to new sso signup command

Problem:

Before the new auth webview, CW used a quickpick wizard for
users to setup their required auth (sso, builderid). They used
a specific command and ensured they used it in their unit tests.

But now with the auth webview changes we will not be redirecting
them to the quickpick wizard anymore. Instead, we will redirect
users to the webview. This uses a different command name.

Solution:

Update the unit tests to point to the new vscode command name
that will be used.

Additionally, since the whole auth webview feature is not ready
we fallback to the old functionality if the user is not in
dev mode (an aws extension setting).

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: change auth id to lowercase

Changes the IDs of service to lowercase
and in some cases certain names should
be one word (CC and CW).

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: lowercase + single service id

converts the service id values to lower case
and makes some as single words

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* uiOpensView: aws.auth.showConnectionsPage > aws.auth.manageConnections

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>