identityCenter: option for credentials to always show form
In the context of the resource explorer auth sign in
we want to allow users to add additional IAM credentials
even if they already have an existing one.

Solution:

In the Credentials form add a mechanism to specify that
we do not care about checking if credentials is currently
connected. This will allow users to continuously add
new credentials to whoever sets the prop checkIfConnected
when creating the component.

Additional:

Fixed identity center not updating since it used the
same mechanism but didn't emit the proper isConnected
status in the background. We still want to emit the
true isConnected status, but in the case of checkIfConnected
being false we don't want to show the user that state.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>