identityCenter: Allow auth setup of existing sso
In the edge case where we set up the resource
explorer with Identity Center, then setup CodeWhisperer
with IdentityCenter we are not able to since the
Identity Center setup for CW sees that the sso
profile already exists.

Solution:

Do not check SSO profile already exists in CW,
instead allow the user to proceed so the existing backend
CW auth can add scopes to the connection.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>