authWebview: Identity Center sign in for Explorer (#3597)
* identityCenter: remove delayed rendering mechanism

Previously this component had a mechanism to
not show its inner content until it was all done
loading.

But from a recent change this is not needed anymore
since it is done in the parent component instead.

Solution:

Remove the old mechanism

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Update some backend methods

The existing identity center methods were
codewhisperer centric.

This commit updates and adds method for general
uses of identity center as well.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Add identity center form state

This adds the class which manages the state of
identity center in the context of the Explorer.

In the context of the Explorer, Identity Center
form behaviour is slightly different which is reflected
in some of the implementations of the class.

In short, IC in Explorer is not required to be
actively connected. Instead it is just used to
get the Credentials profile information. So we
do not care about the active IC connection in this
scenario

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Fix 'waiting for user' stage not being shown

The waiting for user stage is shown once the user starts
sso signup process. It wasn't activated in the actual UI,
this fixes that

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Show different form title for Explorer

With Identity Center in the context of the Explorer
we do not care about the actual 'active' status
of an Identity Center connection.

Due to this we don't want to show anything regarding
a 'successful' connection in the form title.

Solution:

Have a conditional title that simply shows the name
if we set a certain Vue Prop and ignores the connected
stuff.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: partial explorer form setup

I still need to add the extra form that will aggregate
the IC and Credentials status in to one.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: somewhat finished explorer window

Still reloads the whole webview upon submission when
I have an existing successful connection. Due to this
if there is an error I see it briefly before the whole
page is reloaded.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: option for credentials to always show form

In the context of the resource explorer auth sign in
we want to allow users to add additional IAM credentials
even if they already have an existing one.

Solution:

In the Credentials form add a mechanism to specify that
we do not care about checking if credentials is currently
connected. This will allow users to continuously add
new credentials to whoever sets the prop checkIfConnected
when creating the component.

Additional:

Fixed identity center not updating since it used the
same mechanism but didn't emit the proper isConnected
status in the background. We still want to emit the
true isConnected status, but in the case of checkIfConnected
being false we don't want to show the user that state.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: show correct connection name once connected

The ExplorerAggregateForm that represents the status
of all connections relevant to the Explorer will now
show the name of the specific underyling connection
that enables it to be connected.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Force loading of SSO profiles

When we add the new Identity Center/SSO connection
we will now trigger loading of connected Credentials
profiles in to the profile store.

We are doing this so that users of this information
are updated with the new profiles.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>