greenBar: show credentials exist but none are selected
Problem:

In the scenario where a brand new user comes in
with credentials already in their credentials file
we are able to pick those up, but currently don't
indicate this to the user. They must know to open
the connection quick pick and then select the credential
we found.

Solution:
in the auth webview, if we detect credentials and none
are yet active/selected, we will show a status bar
message directing them to select a credentials from
the status bar.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>