authWebview: minor changes (#3599)
* links: Add in missing links

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* ssoError: Show SSO submission error message

No when there is an error with SSO submission,
an error message will appear under the submit
button with an explanation why.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* builderIdError: Show error if builder id setup cancelled

During CW or CC builder id setup the user may cancel
and end the process. We will show an error to indicate
they cancelled under the sign up/sign in button.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* alignForms: Aligns the auth forms to the left

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* codewhisperer: use connection after adding scopes

In the scenario where an sso profile already existed
and we wanted to add codewhisperer scopes to it, we would
add scopes but not use the new sso connection.

This change uses the new connection after adding scopes
to it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Allow auth setup of existing sso

In the edge case where we set up the resource
explorer with Identity Center, then setup CodeWhisperer
with IdentityCenter we are not able to since the
Identity Center setup for CW sees that the sso
profile already exists.

Solution:

Do not check SSO profile already exists in CW,
instead allow the user to proceed so the existing backend
CW auth can add scopes to the connection.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* typo: remove plural

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Update form field subtext

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Edit button to select region

Instead of a dropdown menu UI to select the region, this
replaces it with an edit button to open the region
prompter

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* authWebview: Show on startup w/ option to disable

On startup of the extension we will show the webview, and
in the corner we will show an information message to not show
the window again if they select the option

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* identityCenter: Clear data on successful submission

In a certain scenario when we successfully submit,
then try to add another IC connection we will see the
old data.

This commit clears the old data one a successful submission.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* codewhisperer: refactor sso connection function

We want to call the sso connection function which
connects to the enterprise sso then calls some additional
functionality (refresh and enable code suggestions), but
want to skip the prompt to input the sso data.

Solution:

Extract the sso connection code from getStartUrl()
so we can reuse it in our webview

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* greenBar: show credentials exist but none are selected

Problem:

In the scenario where a brand new user comes in
with credentials already in their credentials file
we are able to pick those up, but currently don't
indicate this to the user. They must know to open
the connection quick pick and then select the credential
we found.

Solution:
in the auth webview, if we detect credentials and none
are yet active/selected, we will show a status bar
message directing them to select a credentials from
the status bar.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* greenBar: quick pick did not select

I chose the wrong prompter which only returned
the connection I chose, not actually use it.

This updates to the correct one.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>