ui: Adjust UI so content window is at top of page
Before we had the content window (with image + auth form)
under the title section so it was adjacent to the service/feature
tab/buttons. But we want to fit everything on one screen and
it required scrolling down a bit.

Solution:

Move the content window to the top of the whole screen
and now it all fits in one screen

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>