bug: Builder ID add scopes fix + missing Event
Problem:

- Sign in to builder id for CW
- Start sign in for builder id for CC
- On the prompt that asks to copy code, click cancel

We are now in a weird state where in the Tree View
we see CW asking for 'Reconnect' but in webview it
is still Enabled.

Solution:

- I was missing an Auth Event to trigger on when
  a connection is updated.
- During the addScopes() process it will update the
  connection, and if we then cancelled the webview
  would be wrong since it only refreshed on a successful
  addScopes() run.
- Now I added in the webview event listener to listen to
  the connection update event and it works

Additional:

If we update scopes of a connection but during reauthenticate()
the user cancels, we are stuck in a weird state where
the scopes on the unauthenticated connection include the
new scopes.

As a fix, we catch if the user cancels and revert the scopes to
what they previously were. The user will still need to reauth
but at least they are shown the actual state of things.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>