refactor: is first extension use logic
Previously we had a flag in settings.json to
indicate if the extension was first used. But
this could be manipulated easily by users.

Solution:

Create a class ExtensionUse that manages this
for you and instead uses the vscode global state
to keep track of this.

Remove the old way we tracked if this was the
first use.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>