refactor: Get CCAuth instance without race condition
Previously the code attempted to handle a race condition
for getting the CC auth instance.

Solution:

Instead we will create the instance if it does not exist
and pass it in as an arg to the class that needed it.

Additionally, remove the old code that waited for the
object to exist since it is not needed anymore.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>