metric: initial UI click infra + use
This sets up the initial ui click infra to call
the metric from the UI and then uses it for a
couple metrics.

Additional:

How the infra works, a component extends the
TelemetryClient vue component and has access to
methods that directly call the ui click backend
command.

We can add more over time to this component.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>