authWebview: Bug Bash fixes (#3620)
* identityCenter: Have button only say 'sign in'

You can't really sign up for builder id

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* authForm: Show error on submit but empty field

If a user clicks the submit button on a form
but they have not entered information in to a
certain field we will now show an errror message
under that field.

Fixes IDE-11014

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* builderId: Submit button text changes on context

If we have a builder id already signed in
on subsequent builder id forms we do not want
to prompt the user to 'sign in' again since they
already are and it may confuse them.

Solution:

If we see builder id is already signed in then
we want to change the text of the submit button
for other builder id forms to have a cta to
'connect builder id' with the other service

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* bug: Builder ID add scopes fix + missing Event

Problem:

- Sign in to builder id for CW
- Start sign in for builder id for CC
- On the prompt that asks to copy code, click cancel

We are now in a weird state where in the Tree View
we see CW asking for 'Reconnect' but in webview it
is still Enabled.

Solution:

- I was missing an Auth Event to trigger on when
  a connection is updated.
- During the addScopes() process it will update the
  connection, and if we then cancelled the webview
  would be wrong since it only refreshed on a successful
  addScopes() run.
- Now I added in the webview event listener to listen to
  the connection update event and it works

Additional:

If we update scopes of a connection but during reauthenticate()
the user cancels, we are stuck in a weird state where
the scopes on the unauthenticated connection include the
new scopes.

As a fix, we catch if the user cancels and revert the scopes to
what they previously were. The user will still need to reauth
but at least they are shown the actual state of things.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: Get CCAuth instance without race condition

Previously the code attempted to handle a race condition
for getting the CC auth instance.

Solution:

Instead we will create the instance if it does not exist
and pass it in as an arg to the class that needed it.

Additionally, remove the old code that waited for the
object to exist since it is not needed anymore.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: Include service names in text content

In the tabs + status bar we will now include more specific
information regarding the service

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* preloadImages: in auth webview

This uses the <img> tag to preload the images
on initial webview load

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: is first extension use logic

Previously we had a flag in settings.json to
indicate if the extension was first used. But
this could be manipulated easily by users.

Solution:

Create a class ExtensionUse that manages this
for you and instead uses the vscode global state
to keep track of this.

Remove the old way we tracked if this was the
first use.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>