error message object for Builder ID + IC
Before we sent a string as the error message.

But we realized that for metrics we will probably
want to know the actual type of error. The issue
is if we use the message as an identifier we cannot
change the message.

Solution:

Send an error object with an id that will remain static
but also a message which can be dynamic

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>