remove choose default region prompt on startup
Previously users would see a quickpick of if they
wanted to add the region that was associated with
their current credentials if it was not already added.

The issue is we don't want to show this quickpick.

Solution:

Don't show the quickpick anymore, just add the region
and if they want they can unadd it in the explorer

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>