fix(ec2): check if instance connectable before connecting.  (#3613)
Problem:
If the user changes the IAM role associated with the instance, the connect command fails whether it be adding or removing the necessary policies.

The root of this problem is ultimately a bug in the usage of the startSession SDK. That is, there are ways this command can fail such that it doesn't throw an error, but the terminal will crash on entry.

Solution:
We check for ways the startSession command can fail before running it. If any are detected, we throw a relevant error. This way, we only call this command when we are confident it will succeed.