ec2: explicitly check if SSM agent is running (#3617)
Problem:
If the error is not related to permissions or instance status, we throw a general error about the SSM agent not running. However, this may not always be the case and we shouldn't assume other errors couldn't occur.

Solution:
We can leverage DescribeInstanceInformation from the SSM SDK to check the ping status of the SSM agent of target EC2 instance. If this call returns an error, or the ping status is outdated, we can conclude the SSM agent isn't running.

Now, there is a specific error message for when an instance is not able to connect. Therefore, the default error message has been changed so that it does not mention the SSM agent (since we would have already checked that).