fix(ec2): instance may not have a name (#3639)
Problem
Currently the toolkit is unable to load EC2 instances that are not assigned a name. It causes it to fail to load all of the instances associated with a specific region.

Solution
displayed as (no name) in QuickPick.
displayed with instanceId in explorer.
behavior above captured in test suite.