authWebview: isConnectionExists method
Adds in a new method for each auth form (eg builder id for CodeWhisperer)
which simply returns a boolean of if a connection for that
form exists (is known by toolkit).

This is being done since we will eventually want to know
for each auth form if a connection exists, since checking
if an auth form connection is actively connected does not
tell us if the connection exists. It may exist but not be
active.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>