refactor: auth webview to use existing auths at start
Previously, we only looked at actively connected auths
when taking the initial snapshot.

Now, we will instead look at all existing auths that
the extension is aware of. This will handle edge cases
where a connection exists, but is not connected and that
does not get considered as an existing connection.

Additionally, renames some variables + minor refactor

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>