fix: disabling credentials submit button appropriately
There was some old logic the enabled/disabled the submit
button. This commit removes it.

Additionally, there was a slight refactor to appropriately
use the submit button disable logic. What it now does is
disables the submit button while the authentication of the
credentials is happening. Once a result is returned, regardless
of success, we enable the button again.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>