bug: ExtensionUserActivity test was flaky
This fixes a flaky test due to the timeout delays
not being properly spaced. Because of this the
vscode Event Emitter would fire in the same
throttle interval. This caused the throttled function to treat
all calls to it to be in the same interval, such
that it only triggered once.

This fixes the delay of when the second set of
event firing goes off so that those are triggered
when we are in the next throttle interval, instead
of the first.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>