feat(ec2): status icons in QuickPick, Explorer (#3661)
Problem

While we have added functionality to start/stop/restart instances from explorer, we don't have anyway of knowing the existing state. 

Solution

First, I want to point out that the icons used here are temporary and I don't believe they are the best fit. They are put there currently to demonstrate the functionality. 

The three different icons to the left of the label indicate the status of the instance with the 'X' meaning stopped, the check mark meaning 'running', and the wheel indicating that its 'pending' (starting up, booting down). 

We also added icons to run/stop/restart the instances, such that the icons are only visible for valid actions (i.e. we cannot stop an already stopped instance). Additionally, we only display the `openTerminal` command next to `running` instances.

Notes on Explorer Updates:

- Currently the command is not accessible from the command palette, because we have no way to link the selection back to the node to update. One approach could be to refresh the entire explorer or find a way to expose the `ec2ParentNode` such that we can refresh just that node. 
- When executing a command, the node will visually update, but only once. Thus, starting a node will leave it stuck in pending status until another refresh from elsewhere allows it to update. This issue will be addressed in a subsequent PR. 
- Explorer updates are limited to the EC2ParentNode, so any actions here will not impact updates for other nodes.