CodeWhisperer: Correct completion type for every completion in user decision events (#3725)
1. Previously we just use the completion type of the first suggestion,
as the completion type for all suggestions. Now since this is no longer true,
plugin side also needs to make the change to compute the completion type
for every suggestion. Since this is per-suggestion level this will
be recorded in the userDecision event.

2. Note that we still have 'completionType' field in other events(
serviceInvocation, clientComponentLatency, etc) these values no longer
make sense for analysis.

3. For actually computing the completionType for each suggestion, the rule
is:
If a suggestion has multi-lines and they are non-blank lines, it's a
block suggestion, otherwise it's a line suggestion.