refactor: create isSsoConnection() alternatives
A sso connection contains different types of sso connections,
such as Builder ID or AWS Identity Center (Base).

This commit creates functions to enable a user to check
with specificity of if a connection is of a specific type
of sso connection.

Additionally, it updates existing uses to the appropriate
implementation.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>