logger: enable logger
This enables the logger, you can now see the logs in
the browser console and the toolkit log file

----

Additional regarding the telemetry stub:

For the logger to run we need to stub the telemetry
client for now since it uses this object and it
must exist.

Eventually we will fully implement telemetry and be
able to remove this.

Signed-off-by: nkomonen <nkomonen@amazon.com>