fix: package.json invalid after 'npm run generatePackage'
This is related to us enabling esModuleInterop in the tsconfig.
When 'npm run generatePackage' was run it would add extra values
to the package.json.

As a solution it was narrowed down that we can not do namespace
imports (import * as assert from 'assert') and must instead do
default imports (import assert from 'assert') for the package.json
file.

Signed-off-by: nkomonen <nkomonen@amazon.com>