refactor: move promisify call inside class
Problem:
When this module is loaded the promisify call gets called
and it fails in the browser.

We are not explicitly using the functionality at the moment.

Solution:
Move the promisify call to inside the class where it is used.
If we do end up using this in the browser we will need to revisit
this.
But for now it allows things to run in the browser

Signed-off-by: nkomonen <nkomonen@amazon.com>