browser: polyfill 'fs' with 'memfs' temporarily
Execution of the extension fails in browser mode
partly due to 'fs' not working in the browser.

Solution:

polyfill 'fs' with 'memfs' TEMPORARILY during the
development phase.

Eventually we will need to see what still depends
on 'fs' and if it is out of our control we may want
to remove our dependency from it.

Signed-off-by: nkomonen <nkomonen@amazon.com>