winston: Fix logging to file in browser
Once I actually ran things in the browser I
found that we didn't have the correct permissions
to write to the log file even though the path
to the log file was provided by vscode itself.

Later I found the reason was that we were not using
vscode.Uri and instead the string path. The key information
that was lost was the file scheme which was hinted at
by vscode#131902 issue.

Solution:

Use vscode.Uri the entire flow for the winston logger.

Signed-off-by: nkomonen <nkomonen@amazon.com>