feature(weaverbird): implement the CodeGenIteration step (#987)
Currently the state was not handling properly iterations in the code generation step. This PR is adding the correct implementation for the code generation iteration step, i.e. when the user adds some comments to the proposed code and/or asks to regenerate the solution to the LLM.

Co-authored-by: Joao Salles <jvsalles@amazon.de>
Co-authored By: Joshua Pinkney <jpink@amazon.com>