sso: use new 'confirm match' device code flow
Problem:

Before this commit, the BuilderID/Identity Center sso flow
would require the user to copy a code and paste it in the
browser window to confirm their device.

Solution:

Now, with new changes from the identity team, if we pass
the device code in the URL a different flow will be shown
in the browser which asks them to instead confirm the
code instead of pasting it.

- This commit has users go through the new flow
- Copying code functionality is removed
- The information messages are updated to mention the requirement
  to confirm the code in the browser

Signed-off-by: nkomonen <nkomonen@amazon.com>